backend/requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
requests>=2.31.0
psycopg2-binary>=2.9.9
networkx>=3.0
matplotlib>=3.8.0
SpeechRecognition>=3.10.0
python-dotenv>=1.0.0
google-generativeai>=0.5.0
# Optional services
bleak>=0.21.0
RPi.GPIO>=0.7.1; platform_system=="Linux"
hx711>=0.5.0; platform_system=="Linux"

